+++ Scripts/Components/Other/DSC_ADD_Velocity.cs
@@ -1,0 +1,11 @@
+﻿using Unity.Entities;
+using Unity.Mathematics;
+
+namespace DSC.Actor.DOTS
+{
+    [GenerateAuthoringComponent]
+    public struct DSC_ADD_Velocity : IComponentData
+    {
+        public float3 m_f3Velocity;
+    }
+}