+++ Scripts/Systems/Actor/GameObject/DSC_ADS_CopyPositionGameObject.cs
@@ -1,0 +1,24 @@
+﻿using UnityEngine;
+using Unity.Entities;
+using Unity.Transforms;
+
+namespace DSC.Actor.DOTS
+{
+    [UpdateInGroup(typeof(DSC_ADG_GameObjectToEntity))]
+    public sealed class DSC_ADS_CopyPositionGameObject : SystemBase
+    {
+        protected override void OnUpdate()
+        {
+            Entities.WithAll<DSC_ADT_CopyPosition>()
+                .ForEach((Transform transform
+                ,ref Translation hTrans) =>
+            {
+                if (transform)
+                {
+                    hTrans.Value = transform.position;
+                }
+            }).WithoutBurst()
+            .Run();
+        }
+    }
+}