+++ Scripts/Systems/Actor/Move/DSC_ADS_FPS_Move_Rigidbody.cs
@@ -1,0 +1,39 @@
+﻿using Unity.Entities;
+using Unity.Mathematics;
+using UnityEngine;
+
+namespace DSC.Actor.DOTS
+{    
+    [UpdateInGroup(typeof(DSC_ADG_Move))]
+    [UpdateAfter(typeof(DSC_ADS_FPS_Move))]
+    public sealed class DSC_ADS_FPS_Move_Rigidbody : SystemBase
+    {
+        protected override void OnUpdate()
+        {
+            float fDeltaTime = Time.DeltaTime;
+
+            var vCamForward = Camera.main.transform.forward;
+            var vCamRight = Camera.main.transform.right;
+
+            Entities.WithAll<DSC_ADT_FPS>()
+                .ForEach((Rigidbody hRigid
+                , ref DSC_ADD_Input hInput
+                , ref DSC_ADD_Move hMove
+                , ref DSC_ADD_MoveSpeed3D hSpeed
+                , ref DSC_ADD_GroundCheck hGround) =>
+            {
+                if (hRigid == null
+                || !hMove.m_bHasMove)
+                    return;
+
+                hRigid.MovePosition(hRigid.position + (Vector3)hMove.m_f3Move);
+
+                hMove.m_bHasMove = false;
+                hMove.m_f3Move = float3.zero;
+
+            }).WithNone<CharacterController>()
+            .WithoutBurst()
+            .Run();
+        }
+    }
+}