+++ Scripts/Systems/Group/DSC_ADG_Move.cs
@@ -1,0 +1,11 @@
+﻿using Unity.Entities;
+using UnityEngine;
+
+namespace DSC.Actor.DOTS
+{
+    [ExecuteAlways]
+    public class DSC_ADG_Move : ComponentSystemGroup
+    {
+
+    }
+}